Fixing bugs on incoming payment request and outgoing payments

+++ mifos-payment/src/main/java/org/mifos/mifospaymentbridge/mifos/domain/savingsaccount/deposit/Summary.java
@@ -14,6 +14,6 @@ public class Summary {
     @JsonProperty("currency")
     public Currency_ currency;
     @JsonProperty("accountBalance")
-    public Integer accountBalance;
+    public Double accountBalance;
 
 }
